# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Robe.AI — Wardrobe Style API",
    "description": "Garment trait extraction, wardrobe management, style profiling, and recommendations.",
    "version": "v1"
  },
  "paths": {
    "/api/garments/analyze": {
      "post": {
        "tags": [
          "Garments"
        ],
        "summary": "Analyze a garment image and extract structured traits.",
        "description": "Stateless — sends the image to a vision model and returns extracted traits.\r\nNo authentication required. No data is persisted.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/garments": {
      "post": {
        "tags": [
          "Garments"
        ],
        "summary": "Store a new garment with its image and extracted traits.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGarmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGarmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGarmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Garments"
        ],
        "summary": "List the current user's garments with optional category filter and paging.",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/GarmentCategory"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarmentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/garments/{id}": {
      "get": {
        "tags": [
          "Garments"
        ],
        "summary": "Get a single garment by ID (must belong to the current user).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GarmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Garments"
        ],
        "summary": "Delete a garment and its stored image (must belong to the current user).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/profile/generate": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "Generate or refresh the current user's style profile from their wardrobe.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get the current user's last generated style profile.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "type": "object",
        "properties": {
          "imageBase64": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AnalyzeResponse": {
        "type": "object",
        "properties": {
          "traits": {
            "$ref": "#/components/schemas/GarmentTraits"
          },
          "modelVersion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ColorInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "hex": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ColorWeight": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "hex": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CreateGarmentRequest": {
        "type": "object",
        "properties": {
          "traits": {
            "$ref": "#/components/schemas/GarmentTraits"
          },
          "imageBase64": {
            "type": "string",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FormalityRange": {
        "type": "object",
        "properties": {
          "min": {
            "type": "integer",
            "format": "int32"
          },
          "max": {
            "type": "integer",
            "format": "int32"
          },
          "typical": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "GarmentCategory": {
        "enum": [
          "top",
          "bottom",
          "dress",
          "outerwear",
          "footwear",
          "accessory",
          "other"
        ],
        "type": "string"
      },
      "GarmentFit": {
        "enum": [
          "slim",
          "regular",
          "loose",
          "oversized"
        ],
        "type": "string"
      },
      "GarmentListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GarmentResponse"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "GarmentPattern": {
        "enum": [
          "solid",
          "striped",
          "plaid",
          "checked",
          "floral",
          "graphic",
          "other"
        ],
        "type": "string"
      },
      "GarmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "traits": {
            "$ref": "#/components/schemas/GarmentTraits"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "modifiedByUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GarmentTraits": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/GarmentCategory"
          },
          "subcategory": {
            "type": "string",
            "nullable": true
          },
          "primaryColor": {
            "$ref": "#/components/schemas/ColorInfo"
          },
          "secondaryColors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorInfo"
            },
            "nullable": true
          },
          "pattern": {
            "$ref": "#/components/schemas/GarmentPattern"
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fit": {
            "$ref": "#/components/schemas/GarmentFit"
          },
          "formality": {
            "type": "integer",
            "format": "int32"
          },
          "seasonality": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Season"
            },
            "nullable": true
          },
          "styleTags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "occasions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "dominantStyles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "colorPalette": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorWeight"
            },
            "nullable": true
          },
          "formalityRange": {
            "$ref": "#/components/schemas/FormalityRange"
          },
          "preferredFits": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "seasonalSkew": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "garmentCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByUserId": {
            "type": "string",
            "nullable": true
          },
          "modifiedByUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Season": {
        "enum": [
          "spring",
          "summer",
          "fall",
          "winter",
          "all"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your Entra ID JWT token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}